requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-stubs==5.2.2
djangorestframework==3.15.2
flake8==7.1.1
flake8-django==1.4
flake8-docstrings==1.7.0
iniconfig==2.0.0
isort==6.0.1
packaging==24.2
Pillow==11.0.0
pluggy==1.5.0
py==1.11.0
pyparsing==3.2.1
pytest==8.3.3
pytest-django==4.9.0
pytz==2024.2
sqlparse==0.5.2
toml==0.10.2